Finish installation components
