+++ src/cli/handlers/uninstall.ts
@@ -2,7 +2,7 @@ import yargs from 'yargs'
 import isAdmin from 'is-admin'
 import deregister from '../../api/deregister'
 import { hasCredential } from '../../common/credentialStore'
-import { uninstallService } from '../../utils/windowsServiceManager'
+import { stopService, uninstallService } from '../../os-shim/windowsServiceDriver'
 import { parseCLIArgs } from '../utils'
 
 export default async (args: yargs.Argv, options?: { failIfAlreadyDone?: boolean, needsArgs?: boolean }): Promise<void> => {
@@ -24,8 +24,9 @@ export default async (args: yargs.Argv, options?: { failIfAlreadyDone?: boolean,
     if (process.env.NODE_ENV === 'development') {
       logger.warn('Running in development mode, not uninstalling service. Please use the deregister command instead.')
     } else {
+      await stopService()
       const wasUninstalled = await uninstallService()
-      // if (options?.failIfAlreadyDone && !wasUninstalled) process.exit(1)
+      if (options?.failIfAlreadyDone && !wasUninstalled) process.exit(1)
     }
   } catch (err) {
     logger.error(`Could not uninstall Windows service: ${err.stack}`)
